move /lib/autosubmit -> /lib/form-view/autosubmit
