+++ lib/form-view/autosubmit/autosubmit.js
@@ -1,5 +1,5 @@
 import o from 'component-dom'
-import request from '../request/request.js'
+import request from '../../request/request.js'
 import serialize from 'get-form-data'
 
 export default function (form, fn, postserialize) {
