+++ lib/form-view/form-view.js
@@ -2,7 +2,7 @@ import o from 'component-dom'
 import spin from 'component-spin'
 import t from 't-component'
 import * as validate from '../validate/validate.js'
-import autosubmit from '../autosubmit/autosubmit.js'
+import autosubmit from './autosubmit/autosubmit'
 import autovalidate from '../autovalidate/autovalidate.js'
 import isMobile from '../is-mobile/is-mobile.js'
 import View from '../view/view.js'
